MTL-2380 Handle cases where the VIP is invalid
Handle cases where there is no network interface with the VIP index.
By default the VIP index is `2`, meaning systems without a `net2` will
end up failing this for-loop.

Regardless of what the VIP index is, ensure we fallback to the
iPXE's natural starting index if the VIP does not exist.