CASMINST-6966 Add `--timeout` to DHCP
Now that MTL-2104 has merged and brought in newer iPXE source code we can finally make use of the `--timeout` parameter when performing DHCP.

This parameter has helped metal-ipxe in the past for troublesome systems such as the Onboard 10GBase-T NICs on mug.

NOTE: This does not change the LINK TIMEOUT, this timeout only comes into effect after a link has UPPED. In other words, if a link does not open then iPXE will move on to the next NIC, if a link does open, iPXE will hangout for the duration of this timeout. This provides a larger window of time for the switches to route packets correctly after the MLAG flips.