CAST-36348: Use bss_script_path instead of debug_script_path
Call the bss_script_path instead of the debug_script_path because the
debug_script_path just drops you into a shell instead of attempting to
boot, which is what we want to debug this issue. Thus, use the
bss_script_path instead because that will cause the node to boot.